refactor: extract processTokenList into a new file
